+++ src/app.module.ts
@@ -4,8 +4,9 @@ import { DataSource } from 'typeorm';
 
 import { Conta } from './domain/entities/conta.entity';
 import { Cliente } from './domain/entities/cliente.entity'; // Importe a entidade Cliente
-import { ContaPoupanca } from './domain/entities/contaPoupanca.entity';
+import { Gerente } from './domain/entities/gerente.entity'; // Importe a entidade Gerente
 import { ContaCorrente } from './domain/entities/contaCorrente.entity';
+import { ContaPoupanca } from './domain/entities/contaPoupanca.entity';
 
 @Module({
   imports: [
@@ -16,7 +17,7 @@ import { ContaCorrente } from './domain/entities/contaCorrente.entity';
       database: 'reprogramabank',
       username: 'reprograma8',
       password: 'repro',
-      entities: [Conta, Cliente],
+      entities: [Cliente, Gerente, Conta, ContaCorrente, ContaPoupanca], 
       synchronize: true, 
     }),
     TypeOrmModule.forFeature([Conta, Cliente]), // Inclua todos os repositórios necessários
