+++ src/domain/entities/cliente.entity.ts
@@ -1,12 +1,35 @@
-import { Entity, PrimaryGeneratedColumn, Column } from 'typeorm';
+import { Entity, PrimaryGeneratedColumn, Column, OneToMany, ManyToOne } from 'typeorm';
+import { v4 as uuidv4 } from 'uuid';
+import { Conta } from './conta.entity';
+import { Gerente } from './gerente.entity';
 
 @Entity()
 export class Cliente {
-  @PrimaryGeneratedColumn()
-  id: number;
+  @PrimaryGeneratedColumn('uuid')
+  id: string;
 
   @Column()
-  nome: string;
+  nomeCompleto: string;
 
-  // Adicione outras propriedades conforme necessário
+  @Column('json')
+  endereco: {
+    rua: string;
+    numero: string;
+    bairro: string;
+    cidade: string;
+    cep: string;
+    estado: string;
+  };
+
+  @Column()
+  telefone: string;
+
+  @OneToMany(() => Conta, (conta) => conta.cliente, { cascade: true })
+  contasAssociadas: Conta[];
+
+  @ManyToOne(() => Gerente, (gerente) => gerente.clientes)
+  gerente: Gerente;
+
+  @Column('decimal')
+  rendaSalarial: number;
 }
