+++ src/domain/entities/contaCorrente.entity.ts
@@ -1,8 +1,16 @@
-// src/entities/conta-corrente.entity.ts
-import { Entity } from 'typeorm';
+import { Entity, Column } from 'typeorm';
 import { Conta } from './conta.entity';
 
 @Entity()
 export class ContaCorrente extends Conta {
-  // Adicione propriedades e métodos específicos para Conta Corrente, se necessário
+  @Column('decimal')
+  limiteChequeEspecial: number;
+
+  usarChequeEspecial(valor: number): boolean {
+    if (this.saldo + this.limiteChequeEspecial >= valor) {
+      this.saldo -= valor;
+      return true;
+    }
+    return false;
+  }
 }
