+++ src/domain/entities/contaPoupanca.entity.ts
@@ -1,8 +1,8 @@
-// src/entities/conta-poupanca.entity.ts
-import { Entity } from 'typeorm';
+import { Entity, Column } from 'typeorm';
 import { Conta } from './conta.entity';
 
 @Entity()
 export class ContaPoupanca extends Conta {
-  // Adicione propriedades e métodos específicos para Conta Poupança, se necessário
+  @Column('decimal')
+  taxaJuros: number;
 }
